refactor(agent-service): dedupe result row formatting
