+++ agent-service/src/agent/tools/result-formatting.ts
@@ -20,6 +20,7 @@
 import { OperatorResultMode, type OperatorExecutionSummary, type SampleRow } from "../../types/execution";
 import {
   formatExecuteOperatorResult,
+  formatSampleRowsAsTsv,
   getOperatorErrorText,
   getOperatorWarnings,
   getVisibleResultHeaders,
@@ -54,7 +55,7 @@ export function formatOperatorResult(operatorId: string, opInfo: OperatorExecuti
   const headers = rows.length > 0 ? getVisibleResultHeaders(rows[0].tuple) : [];
   const columns = headers.length;
 
-  const dataString = jsonToTableFormat(rows);
+  const dataString = formatSampleRowsAsTsv(rows);
 
   // Output shape only; input-port shapes are derivable by the agent from the DAG
   // links plus each upstream operator's own output shape shown in context.
@@ -65,38 +66,4 @@ export function formatOperatorResult(operatorId: string, opInfo: OperatorExecuti
 
   const briefSummary = formatExecuteOperatorResult(operatorId);
   return [briefSummary, ...metadataLines, dataString].filter(Boolean).join("\n");
-}
-
-function jsonToTableFormat(rows: SampleRow[]): string {
-  if (!rows || rows.length === 0) return "";
-
-  const headers = getVisibleResultHeaders(rows[0].tuple);
-  if (headers.length === 0) return "";
-
-  const headerLine = "\t" + headers.join("\t");
-  const formattedRows: string[] = [];
-  let prevIndex = -1;
-
-  for (const { rowIndex, tuple } of rows) {
-    if (prevIndex >= 0 && rowIndex > prevIndex + 1) {
-      const dots = headers.map(() => "...").join("\t");
-      formattedRows.push(`...\t${dots}`);
-    }
-    prevIndex = rowIndex;
-
-    const cells = headers.map(h => {
-      const val = tuple[h];
-      if (val === null) return "NaN";
-      if (val === undefined) return "";
-      if (typeof val === "number" || typeof val === "boolean") return String(val);
-      if (typeof val === "string") {
-        if (val === "NULL") return "NaN";
-        return val.replace(/\t/g, "\\t").replace(/\n/g, "\\n");
-      }
-      return JSON.stringify(val);
-    });
-    formattedRows.push(`${rowIndex}\t${cells.join("\t")}`);
-  }
-
-  return [headerLine, ...formattedRows].join("\n");
 }
