+++ frontend/src/app/workspace/component/result-panel/result-table-frame/result-table-frame.component.spec.ts
@@ -73,14 +73,4 @@ describe("ResultTableFrameComponent", () => {
   it("should set columnLimit from gui-config", () => {
     expect(component.columnLimit).toEqual(GUI_CONFIG_LIMIT);
   });
-
-  it("populates currentResult, columns, and totalNumTuples when given non-empty data", () => {
-    component.operatorId = "op-1";
-
-    (component as any).setupResultTable([{ colA: "x", colB: 1 }], 42);
-
-    expect(component.currentResult).toEqual([{ colA: "x", colB: 1 }]);
-    expect(component.totalNumTuples).toEqual(42);
-    expect(component.currentColumns!.length).toBeGreaterThan(0);
-  });
 });
